Remove support for disabling the non-PKCE OIDC authentication flow
This commit removes the original OIDC Code Redirect authentication workflow in support of the PKCE
enabled workflow. This code was originally gated to enable backward compatibility.
