+++ app/db/repository/authenticator_repository.rb
@@ -4,13 +4,14 @@ module DB
       def initialize(
         data_object:,
         resource_repository: ::Resource,
-        logger: Rails.logger,
-        pkce_support_enabled: Rails.configuration.feature_flags.enabled?(:pkce_support)
+        logger: Rails.logger
+        # ,
+        # pkce_support_enabled: Rails.configuration.feature_flags.enabled?(:pkce_support)
       )
         @resource_repository = resource_repository
         @data_object = data_object
         @logger = logger
-        @pkce_support_enabled = pkce_support_enabled
+        # @pkce_support_enabled = pkce_support_enabled
       end
 
       def find_all(type:, account:)
@@ -73,12 +74,10 @@ module DB
         end
 
         begin
-          if @pkce_support_enabled
-            allowed_args = %i[account service_id] +
-                          @data_object.const_get(:REQUIRED_VARIABLES) +
-                          @data_object.const_get(:OPTIONAL_VARIABLES)
-            args_list = args_list.select{ |key, value| allowed_args.include?(key) && value.present? }
-          end
+          allowed_args = %i[account service_id] +
+                        @data_object.const_get(:REQUIRED_VARIABLES) +
+                        @data_object.const_get(:OPTIONAL_VARIABLES)
+          args_list = args_list.select { |key, value| allowed_args.include?(key) && value.present? }
           @data_object.new(**args_list)
         rescue ArgumentError => e
           @logger.debug("DB::Repository::AuthenticatorRepository.load_authenticator - exception: #{e}")
