+++ app/domain/authentication/authn_oidc/v2/data_objects/authenticator.rb
@@ -4,17 +4,25 @@ module Authentication
       module DataObjects
         class Authenticator
 
-          # required
-          attr_reader :provider_uri, :client_id, :client_secret, :claim_mapping, :nonce, :state, :account
-          attr_reader :service_id, :redirect_uri, :response_type
+          REQUIRED_VARIABLES = %i[provider_uri client_id client_secret claim_mapping].freeze
+          OPTIONAL_VARIABLES = %i[redirect_uri response_type provider_scope name token_ttl].freeze
+
+          attr_reader(
+            :provider_uri,
+            :client_id,
+            :client_secret,
+            :claim_mapping,
+            :account,
+            :service_id,
+            :redirect_uri,
+            :response_type
+          )
 
           def initialize(
             provider_uri:,
             client_id:,
             client_secret:,
             claim_mapping:,
-            nonce:,
-            state:,
             account:,
             service_id:,
             redirect_uri: nil,
@@ -28,9 +36,7 @@ module Authentication
             @client_id = client_id
             @client_secret = client_secret
             @claim_mapping = claim_mapping
-            @nonce = nonce
             @response_type = response_type
-            @state = state
             @service_id = service_id
             @name = name
             @provider_scope = provider_scope
@@ -39,7 +45,7 @@ module Authentication
           end
 
           def scope
-            (%w[openid email profile] + [*@provider_scope]).uniq.join(' ')
+            (%w[openid email profile] + [*@provider_scope.to_s.split(' ')]).uniq.join(' ')
           end
 
           def name
