+++ spec/app/domain/authentication/authn-oidc/v2/client_spec.rb
@@ -10,98 +10,129 @@ RSpec.describe(Authentication::AuthnOidc::V2::Client) do
       client_id: '0oa3w3xig6rHiu9yT5d7',
       client_secret: 'e349BMTTIpLO-rPuPqLLkLyH_pO-loUzhIVJCrHj',
       claim_mapping: 'foo',
-      nonce: '1656b4264b60af659fce',
-      state: 'state',
       account: 'bar',
       service_id: 'baz'
     )
   end
 
   let(:client) do
-    Authentication::AuthnOidc::V2::Client.new(
-      authenticator: authenticator
-    )
+    VCR.use_cassette("authenticators/authn-oidc/pkce_support_feature/client_load") do
+      client = Authentication::AuthnOidc::V2::Client.new(
+        authenticator: authenticator
+      )
+      # The call `oidc_client` queries the OIDC endpoint. As such,
+      # we need to wrap this in a VCR call. Calling this before
+      # returning the client to allow this call to be more effectively
+      # mocked.
+      client.oidc_client
+      client
+    end
   end
 
   describe '.callback' do
     context 'when credentials are valid' do
-      it 'returns a valid JWT token', vcr: 'authenticators/authn-oidc/v2/client_callback-valid_oidc_credentials' do
+      it 'returns a valid JWT token', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-valid_oidc_credentials' do
         # Because JWT tokens have an expiration timeframe, we need to hold
         # time constant after caching the request.
-        travel_to(Time.parse("2022-06-30 16:42:17 +0000")) do
+        travel_to(Time.parse("2022-09-30 17:02:17 +0000")) do
           token = client.callback(
-            code: 'qdDm7On1dEEzNmMlk2bF7IcOF8gCgfvgMCMXXXDlYEE'
+            code: '-QGREc_SONbbJIKdbpyYudA13c9PZlgqdxowkf45LOw',
+            code_verifier: 'c1de7f1251849accd99d4839d79a637561b1181b909ed7dc1d',
+            nonce: '7efcbba36a9b96fdb5285a159665c3d382abd8b6b3288fcc8d'
           )
           expect(token).to be_a_kind_of(OpenIDConnect::ResponseObject::IdToken)
-          expect(token.raw_attributes['nonce']).to eq('1656b4264b60af659fce')
+          expect(token.raw_attributes['nonce']).to eq('7efcbba36a9b96fdb5285a159665c3d382abd8b6b3288fcc8d')
           expect(token.raw_attributes['preferred_username']).to eq('test.user3@mycompany.com')
           expect(token.aud).to eq('0oa3w3xig6rHiu9yT5d7')
         end
       end
     end
 
-    context 'when JWT has expired' do
-      it 'raises an error', vcr: 'authenticators/authn-oidc/v2/client_callback-valid_oidc_credentials' do
-        travel_to(Time.parse("2022-06-30 20:42:17 +0000")) do
+    context 'when code verifier does not match' do
+      it 'raises an error', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-invalid_code_verifier' do
+        travel_to(Time.parse("2022-10-17 17:23:30 +0000")) do
           expect do
             client.callback(
-              code: 'qdDm7On1dEEzNmMlk2bF7IcOF8gCgfvgMCMXXXDlYEE'
+              code: 'GV48_SF4a19ghvBhVbbSG3Lr8BuFl8PhWVPZSbokV2o',
+              code_verifier: 'bad-code-verifier',
+              nonce: '3e6bd5235e4692b37ca1f04cb01b6e0cb177aa20dcef19e89f'
             )
           end.to raise_error(
-            OpenIDConnect::ResponseObject::IdToken::ExpiredToken,
-            'Invalid ID token: Expired token'
+            Errors::Authentication::AuthnOidc::TokenRetrievalFailed,
+            "CONJ00133E Access Token retrieval failure: 'PKCE verification failed'"
+          )
+        end
+      end
+    end
+
+    context 'when nonce does not match' do
+      it 'raises an error', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-valid_oidc_credentials' do
+        travel_to(Time.parse("2022-09-30 17:02:17 +0000")) do
+          expect do
+            client.callback(
+              code: '-QGREc_SONbbJIKdbpyYudA13c9PZlgqdxowkf45LOw',
+              code_verifier: 'c1de7f1251849accd99d4839d79a637561b1181b909ed7dc1d',
+              nonce: 'bad-nonce'
+            )
+          end.to raise_error(
+            Errors::Authentication::AuthnOidc::TokenVerificationFailed,
+            "CONJ00128E JWT Token validation failed: 'Provided nonce does not match the nonce in the JWT'"
+          )
+        end
+      end
+    end
+
+    context 'when JWT has expired' do
+      it 'raises an error', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-valid_oidc_credentials' do
+        travel_to(Time.parse("2022-10-01 17:02:17 +0000")) do
+          expect do
+            client.callback(
+              code: '-QGREc_SONbbJIKdbpyYudA13c9PZlgqdxowkf45LOw',
+              code_verifier: 'c1de7f1251849accd99d4839d79a637561b1181b909ed7dc1d',
+              nonce: '7efcbba36a9b96fdb5285a159665c3d382abd8b6b3288fcc8d'
+            )
+          end.to raise_error(
+            Errors::Authentication::AuthnOidc::TokenVerificationFailed,
+            "CONJ00128E JWT Token validation failed: 'JWT has expired'"
           )
         end
       end
     end
 
     context 'when code has previously been used' do
-      it 'raise an exception', vcr: 'authenticators/authn-oidc/v2/client_callback-used_code-valid_oidc_credentials' do
+      it 'raise an exception', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-used_code-valid_oidc_credentials' do
         expect do
           client.callback(
-            code: '7wKEGhsN9UEL5MG9EfDJ8KWMToKINzvV29uyPsQZYpo'
+            code: '-QGREc_SONbbJIKdbpyYudA13c9PZlgqdxowkf45LOw',
+            code_verifier: 'c1de7f1251849accd99d4839d79a637561b1181b909ed7dc1d',
+            nonce: '7efcbba36a9b96fdb5285a159665c3d382abd8b6b3288fcc8d'
           )
         end.to raise_error(
-          Rack::OAuth2::Client::Error,
-          'invalid_grant :: The authorization code is invalid or has expired.'
+          Errors::Authentication::AuthnOidc::TokenRetrievalFailed,
+          "CONJ00133E Access Token retrieval failure: 'Authorization code is invalid or has expired'"
         )
       end
     end
 
-    context 'when code has expired', vcr: 'authenticators/authn-oidc/v2/client_callback-expired_code-valid_oidc_credentials' do
+    context 'when code has expired', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_callback-expired_code-valid_oidc_credentials' do
       it 'raise an exception' do
         expect do
           client.callback(
-            code: 'SNSPeiQJ0-D6nUHTg-Ht9ZoDxIaaWBB80pnYuXY2VxU'
+            code: 'SNSPeiQJ0-D6nUHTg-Ht9ZoDxIaaWBB80pnYuXY2VxU',
+            code_verifier: 'c1de7f1251849accd99d4839d79a637561b1181b909ed7dc1d',
+            nonce: '7efcbba36a9b96fdb5285a159665c3d382abd8b6b3288fcc8d'
           )
         end.to raise_error(
-          Rack::OAuth2::Client::Error,
-          'invalid_grant :: The authorization code is invalid or has expired.'
+          Errors::Authentication::AuthnOidc::TokenRetrievalFailed,
+          "CONJ00133E Access Token retrieval failure: 'Authorization code is invalid or has expired'"
         )
-      end
-    end
-
-    context 'when code is invalid' do
-      context 'raise an error when' do
-        it 'code is nil' do
-          expect { client.callback(code: nil) }.to raise_error(
-            Errors::Authentication::RequestBody::MissingRequestParam,
-            "CONJ00009E Field 'code' is missing or empty in request body"
-          )
-        end
-        it 'code is an empty string' do
-          expect { client.callback(code: '') }.to raise_error(
-            Errors::Authentication::RequestBody::MissingRequestParam,
-            "CONJ00009E Field 'code' is missing or empty in request body"
-          )
-        end
       end
     end
   end
 
   describe '.oidc_client' do
     context 'when credentials are valid' do
-      it 'returns a valid oidc client', vcr: 'authenticators/authn-oidc/v2/client_initialization' do
+      it 'returns a valid oidc client', vcr: 'authenticators/authn-oidc/pkce_support_feature/client_initialization' do
         oidc_client = client.oidc_client
 
         expect(oidc_client).to be_a_kind_of(OpenIDConnect::Client)
@@ -118,7 +149,7 @@ RSpec.describe(Authentication::AuthnOidc::V2::Client) do
     end
   end
 
-  describe '.discovery_information', vcr: 'authenticators/authn-oidc/v2/discovery_endpoint-valid_oidc_credentials' do
+  describe '.discovery_information', vcr: 'authenticators/authn-oidc/pkce_support_feature/discovery_endpoint-valid_oidc_credentials' do
     context 'when credentials are valid' do
       it 'endpoint returns valid data' do
         discovery_information = client.discovery_information(invalidate: true)
@@ -141,7 +172,7 @@ RSpec.describe(Authentication::AuthnOidc::V2::Client) do
       end
     end
 
-    context 'when provider URI is invalid', vcr: 'authenticators/authn-oidc/v2/discovery_endpoint-invalid_oidc_provider' do
+    context 'when provider URI is invalid', vcr: 'authenticators/authn-oidc/pkce_support_feature/discovery_endpoint-invalid_oidc_provider' do
       it 'returns an timeout error' do
         client = Authentication::AuthnOidc::V2::Client.new(
           authenticator: Authentication::AuthnOidc::V2::DataObjects::Authenticator.new(
@@ -150,8 +181,6 @@ RSpec.describe(Authentication::AuthnOidc::V2::Client) do
             client_id: '0oa3w3xig6rHiu9yT5d7',
             client_secret: 'e349BMTTIpLO-rPuPqLLkLyH_pO-loUzhIVJCrHj',
             claim_mapping: 'foo',
-            nonce: '1656b4264b60af659fce',
-            state: 'state',
             account: 'bar',
             service_id: 'baz'
           )
