+++ spec/app/domain/authentication/authn-oidc/v2/data_objects/authenticator.rb
@@ -9,8 +9,6 @@ RSpec.describe(Authentication::AuthnOidc::V2::DataObjects::Authenticator) do
       client_id: 'client-id-123',
       client_secret: 'client-secret-123',
       claim_mapping: 'email',
-      nonce: 'nonce456',
-      state: 'state123',
       account: 'default',
       service_id: 'my-authenticator'
     }
@@ -34,14 +32,19 @@ RSpec.describe(Authentication::AuthnOidc::V2::DataObjects::Authenticator) do
       it { expect(authenticator.scope).to eq('openid email profile 1') }
     end
 
+    context 'when initialized with a duplicated argument' do
+      let(:args) { default_args.merge({ provider_scope: 'profile' }) }
+      it { expect(authenticator.scope).to eq('openid email profile') }
+    end
+
     context 'when initialized with an array argument' do
       context 'single value array' do
-        let(:args) { default_args.merge({ provider_scope: ['foo'] }) }
+        let(:args) { default_args.merge({ provider_scope: 'foo' }) }
         it { expect(authenticator.scope).to eq('openid email profile foo') }
       end
 
       context 'multi-value array' do
-        let(:args) { default_args.merge({ provider_scope: ['foo', 'bar'] }) }
+        let(:args) { default_args.merge({ provider_scope: 'foo bar' }) }
         it { expect(authenticator.scope).to eq('openid email profile foo bar') }
       end
     end
