+++ spec/app/domain/authentication/authn-oidc/v2/strategy_rspec.rb
@@ -3,7 +3,6 @@
 require 'spec_helper'
 
 RSpec.describe(' Authentication::AuthnOidc::V2::Strategy') do
-
   let(:jwt) { double(raw_attributes: { claim_mapping: "alice" }) }
 
   let(:authenticator) do
@@ -13,18 +12,15 @@ RSpec.describe(' Authentication::AuthnOidc::V2::Strategy') do
       redirect_uri: "http://conjur/authn-oidc/cucumber/authenticate",
       provider_uri: "http://test",
       name: "foo",
-      state: 'foostate',
       client_id: "ConjurClient",
       client_secret: 'client_secret',
-      claim_mapping: mapping,
-      nonce: 'secret'
+      claim_mapping: mapping
     )
   end
 
   let(:client) do
     class_double(::Authentication::AuthnOidc::V2::Client).tap do |double|
-      allow(double).to receive(:new).and_return(
-        current_client)
+      allow(double).to receive(:new).and_return(current_client)
     end
   end
 
@@ -45,24 +41,32 @@ RSpec.describe(' Authentication::AuthnOidc::V2::Strategy') do
     context 'when a role_id matches the identity exist' do
       let(:mapping) { "claim_mapping" }
       it 'returns the role' do
-        expect(strategy.callback({ state: "foostate", code: "code" }))
-          .to eq("alice")
+        expect(strategy.callback(nonce: 'nonce', code: 'code', code_verifier: 'foo'))
+          .to eq('alice')
       end
     end
 
-    context 'When the authn state doesnt match the arguments' do
-      let(:mapping) { "claim_mapping" }
-      it 'raises an error' do
-        expect { strategy.callback({ state: "barstate", code: "code" }) }
-          .to raise_error(Errors::Authentication::AuthnOidc::StateMismatch)
-      end
-    end
-
-    context 'When the claiming matching in the token doesnt match the jwt' do
-      let(:mapping) { "wrong_mapping" }
-      it 'raises an error' do
-        expect { strategy.callback({ state: "foostate", code: "code" }) }
+    context "when the claiming matching in the token doesn't match the jwt" do
+      let(:mapping) { 'wrong_mapping' }
+      it 'raises a `IdTokenClaimNotFoundOrEmpty` error' do
+        expect { strategy.callback(code: 'code', nonce: 'nonce', code_verifier: 'foo') }
           .to raise_error(Errors::Authentication::AuthnOidc::IdTokenClaimNotFoundOrEmpty)
+      end
+    end
+
+    context 'when required parameters are missing' do
+      let(:mapping) { '' }
+      context 'when code is missing' do
+        it 'raises a `MissingRequestParam` error' do
+          expect { strategy.callback(nonce: 'nonce', code_verifier: 'foo') }
+            .to raise_error(Errors::Authentication::RequestBody::MissingRequestParam)
+        end
+      end
+      context 'when nonce is missing' do
+        it 'raises a `MissingRequestParam` error' do
+          expect { strategy.callback(code: 'code', code_verifier: 'foo') }
+            .to raise_error(Errors::Authentication::RequestBody::MissingRequestParam)
+        end
       end
     end
   end
